phone changed to big int

+++ db/migrate/20140730151834_alter_users.rb
@@ -1,6 +1,6 @@
 class AlterUsers < ActiveRecord::Migration
   def up
-    add_column("users", "phone", :integer, :limit => 15, :after => "email")
+    add_column("users", "phone", :bigint, :limit => 15, :after => "email")
     add_column("users", "address", :string, :limit => 150, :after => "phone")
   end
